+++ jdaviz/configs/default/plugins/plot_options/plot_options.vue
@@ -186,6 +186,19 @@
       <glue-float-field label="Stretch VMax" :value.sync="stretch_vmax_value" />
     </glue-state-sync-wrapper>
 
+    <v-row v-if="stretch_function_sync.in_subscribed_states">
+      <!-- z-index to ensure on top of the jupyter widget with negative margin-top -->
+      <v-switch
+        v-model="stretch_hist_zoom_limits"
+        class="hide-input"
+        label="Limit histogram to current zoom limits"
+        style="z-index: 1"
+      ></v-switch>
+      <!-- NOTE: height defined here should match that in the custom CSS rules
+           below for the bqplot class -->
+      <jupyter-widget :widget="stretch_histogram" class="stretch-hist" style="width: 100%; height: 320px; margin-top: -60px; margin-bottom: -40px" />
+    </v-row>
+
     <!-- IMAGE:IMAGE -->
     <j-plugin-section-header v-if="image_visible_sync.in_subscribed_states">Image</j-plugin-section-header>
     <glue-state-sync-wrapper :sync="image_visible_sync" :multiselect="multiselect" @unmix-state="unmix_state('image_visible')">
@@ -385,10 +398,13 @@ module.exports = {
 }
 </script>
 
-<style>
+<style scoped>
 .color-menu {
     font-size: 16px;
     padding-left: 16px;
     border: 2px solid rgba(0,0,0,0.54);
+}
+.stretch-hist > .bqplot {
+  height: 320px !important;
 }
 </style>